Module collenchyma::framework

Provides the generic functionality of a hardware supporting frameworks such as native CPU, OpenCL, CUDA, etc..

The default Framework would be plain host CPU for common computation. To make use of other computation hardwares such as GPUs you would choose other computation Frameworks such as OpenCL or CUDA, which provide the access to those hardwares for computation.

To start backend-agnostic and highly parallel computation, you start by initializing on of the Framework implementations, resulting in an initialized Framework, that contains among other things, a list of all available hardwares through that framework.

Examples

// Initializing a Framework
// let framework: Framework = OpenCL::new();
// let backend: Backend = framework.create_backend();

Enums

Error
Defines a generic set of Framework Errors.

Traits

IFramework
Defines a Framework.