Module collenchyma::frameworks::cuda

Provides informations about the software system, such as Cuda, CUDA, that contains the set of components to support hardwares with kernel execution.

Ontology

Cuda device -> Hardware Cuda context -> Device

Reexports

pub use self::memory::Memory;

pub use self::context::Context;

pub use self::function::Function;

pub use self::module::Module;

pub use self::device::{Device, DeviceInfo};

Modules

context
Provides a Rust wrapper around Cuda's context.

device
Provides a Rust wrapper around Cuda's device.

function
Provides a Rust wrapper around Cuda's Function.

memory

module
Provides a Rust wrapper around Cuda's Module.

Structs

API
Defines the Cuda API.

Cuda
Provides the Cuda Framework.

Enums

Error
Defines OpenCL errors.