Struct collenchyma::frameworks::cuda::device::Device

pub struct Device {
    // some fields omitted
}

Defines a Cuda Device.

Methods

impl Device

fn from_isize(id: isize) -> Device

Initializes a new Cuda device.

fn from_c(id: device_id) -> Device

Initializes a new Cuda device from its C type.

fn id_c(&self) -> device_id

Returns the id as its C type.

fn load_name(&mut self) -> Self

Loads the name of the device via a foreign Cuda call.

fn load_device_type(&mut self) -> Self

Loads the device type via a foreign Cuda call.

fn load_compute_units(&mut self) -> Self

Loads the compute units of the device via a foreign Cuda call.