Struct collenchyma::frameworks::cuda::API

pub struct API;

Defines the Cuda API.

Methods

impl API

fn create_context(devices: Vec<Device>, properties: *const context_properties, callback: extern "C" fn(*const c_char, *const c_void, size_t, *mut c_void), user_data: *mut c_void) -> Result<context_id, Error>

Creates a OpenCL context.

An OpenCL context is created with one or more devices. Contexts are used by the OpenCL runtime for managing objects such as command-queues, memory, program and kernel objects and for executing kernels on one or more devices specified in the context. An OpenCL context is a synonym to a Collenchyma device.

impl API

fn load_devices() -> Result<Vec<Device>, Error>

Returns fully initialized devices for a specific platform.

Combines the fetching of all device ids and the fetching of the individual device information.

fn load_device_list() -> Result<Vec<Device>, Error>

Returns a list of available devices for the provided platform.

fn load_device_info(device: &Device, info: device_info) -> Result<DeviceInfo, Error>

Returns the requested DeviceInfo for the provided device.

impl API

fn create_buffer(context: Context) -> Result<memory_id, Error>

Allocates memory on the OpenCL device.

A buffer object stores a one-dimensional collection of elements. Elements of a buffer object can be a scalar data type (such as an int, float), vector data type, or a user-defined structure. Returns a memory id for the created buffer, which can now be writen to.

fn release_memory(memory: &mut Memory) -> Result<(), Error>

Releases allocated memory from the OpenCL device.

fn read_from_buffer<T>(mem: Memory, blocking_read: boolean, offset: size_t, size: size_t, ptr: *mut c_void, num_events_in_wait_list: uint, event_wait_list: *const event, event: *mut event) -> Result<(), Error>

Reads from a buffer to the host memory.

With write_to_buffer you can do the opposite, write from the host memory to a buffer.