Module collenchyma::frameworks::native

Provides informations about the software system, such as OpenCL, CUDA, that contains the set of components to support hardwares with kernel execution.

Reexports

pub use self::device::Cpu;

pub use self::function::Function;

pub use self::binary::Binary;

Modules

binary
Provides a binary on native CPU.

device
Provides a hardware aka. the host CPU.

flatbox
Provides a Box without any knowledge of its underlying type.

function
Provides a operation on native CPU.

hardware
Provides a hardware aka. the host CPU.

Structs

Native
Provides the Native framework.

Enums

Error
Defines the Native Error.

Traits

INative
Provides the Native framework trait for explicit Backend behaviour.

Functions

allocate_boxed_slice
Traditional allocation via creation of a filled Vec.