Module collenchyma::frameworks::opencl

Provides informations about the software system, such as OpenCL, CUDA, that contains the set of components to support hardwares with kernel execution.

Ontology

OpenCL device -> Hardware OpenCL context -> Device

Reexports

pub use self::platform::Platform;

pub use self::context::Context;

pub use self::memory::Memory;

pub use self::queue::Queue;

pub use self::kernel::Kernel;

pub use self::program::Program;

pub use self::device::{Device, DeviceInfo};

Modules

context
Provides a Rust wrapper around OpenCL's context.

device
Provides a Rust wrapper around OpenCL's device.

kernel
Provides a Rust wrapper around OpenCL's Kernel.

libraries
Provides support for various libraries for a OpenCL backend.

memory

platform
Provides a Rust wrapper around OpenCL's platform.

program
Provides a Rust wrapper around OpenCL's Program.

queue
Provides a Rust wrapper around OpenCL's command queue.

Structs

API
Defines the OpenCL API.

OpenCL
Provides the OpenCL Framework.

Enums

Error
Defines OpenCL errors.

Traits

IOpenCL
Provides the OpenCL framework trait for explicit Backend behaviour.