Module collenchyma::frameworks::opencl::queue

Provides a Rust wrapper around OpenCL's command queue.

OpenCL Command Queue

OpenCL objects such as memory, program and kernel objects are created using a context. Operations on these objects are performed using a command-queue. The command-queue can be used to queue a set of operations (referred to as commands) in order. Having multiple command-queues allows applications to queue multiple independent commands without requiring synchronization. Note that this should work as long as these objects are not being shared. Sharing of objects across multiple command-queues will require the application to perform appropriate synchronization.

Structs

Queue
Defines a OpenCL Queue.