Module collenchyma::hardware

Provides a representation for a collection of available compute units e.g. CPUs or GPUs.

Hardware can be GPUs, multi-core CPUs or DSPs, Cell/B.E. processor or whatever else is supported by the provided frameworks such as OpenCL, CUDA, etc. The struct holds all important information about the hardware. To execute code on hardware, turn hardware into a device.

Enums

HardwareType
Specifies the available Hardware types.

Traits

IHardware
Specifies Hardware behavior accross frameworks.