Module collenchyma::libraries

Exposes the explicit implementations of operations on a backend.

A library can be directly implemented for a backend. A library provides 'provided methods', meaning, that no code needs to be writen at the implementation of the library on a backend.

A library plays a significant role in Collenchyma and allows for the native interface for executing operations, which is no-different than calling a 'normal' Rust method. The library declares the functionality and automatically manages the shared memory arguments (synchronizations and memory creations).

Examples of libraries would be BLAS or a special purpose Neural Network library like cuDNN. But unlike BLAS or cuDNN, a Collenchyma library is completely backend-agnostic and can operate on any framework-supported hardware.

Collenchyma ships with the most basic operations, but you should be able to easily write your own backend-agnostic operations, too.

Modules

blas
Provides backend-agnostic BLAS operations.

numeric_helpers
Describes the Library numeric types and traits.

Enums

Error
Defines a high-level library Error.

Traits

Float