Module collenchyma::libraries::blas

Provides backend-agnostic BLAS operations.

BLAS (Basic Linear Algebra Subprograms) is a specification that prescribes a set of low-level routines for performing common linear algebra operations such as vector addition, scalar multiplication, dot products, linear combinations, and matrix multiplication. They are the de facto standard low-level routines for linear algebra libraries; the routines have bindings for both C and Fortran. Although the BLAS specification is general, BLAS implementations are often optimized for speed on a particular machine, so using them can bring substantial performance benefits. BLAS implementations will take advantage of special floating point hardware such as vector registers or SIMD instructions.
Source

Enums

Error
Defines Blas Errors.

Traits

IBlas
Provides the functionality for a backend to support Basic Linear Algebra Subprogram operations.

IBlasBinary
Describes the operation binding for a Blas Binary implementation.

IOperationAsum
Describes a Asum Operation.

IOperationAxpy
Describes a Axpy Operation.

IOperationCopy
Describes a Copy Operation.

IOperationDot
Describes a Dot Operation.

IOperationNrm2
Describes a Nrm2 Operation.

IOperationScale
Describes a Scale Operation.

IOperationSwap
Describes a Swap Operation.