Trait collenchyma::libraries::blas::IBlas

pub trait IBlas<F: Float> {
    type B: IBlasBinary<F> + IBinary;
    fn binary(&self) -> Self::B;
    fn device(&self) -> &DeviceType;

    fn asum(&self, x: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn axpy(&self, a: &mut SharedMemory<F>, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn copy(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn dot(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn nrm2(&self, x: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn scale(&self, a: &mut SharedMemory<F>, x: &mut SharedMemory<F>) -> Result<(), Error> { ... }
    fn swap(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error> { ... }
}

Provides the functionality for a backend to support Basic Linear Algebra Subprogram operations.

Associated Types

type B: IBlasBinary<F> + IBinary

The Binary representation for this Library.

Required Methods

fn binary(&self) -> Self::B

Returns the binary representation

fn device(&self) -> &DeviceType

Returns the device representation

Provided Methods

fn asum(&self, x: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error>

Computes the absolute sum of vector x.

Saves the result to result. This is a Level 1 BLAS operation.

fn axpy(&self, a: &mut SharedMemory<F>, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error>

Computes a vector x times a constant a plus a vector y aka. a * x + y.

Saves the resulting vector back into y. This is a Level 1 BLAS operation.

fn copy(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error>

Copies x.len() elements of vector x into vector y.

Saves the result to y. This is a Level 1 BLAS operation.

fn dot(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error>

Computes the dot product over x and y.

Saves the resulting value into result. This is a Level 1 BLAS operation.

fn nrm2(&self, x: &mut SharedMemory<F>, result: &mut SharedMemory<F>) -> Result<(), Error>

Computes the L2 norm aka. euclidean length of vector x.

Saves the result to result. This is a Level 1 BLAS operation.

fn scale(&self, a: &mut SharedMemory<F>, x: &mut SharedMemory<F>) -> Result<(), Error>

Scales a vector x by a constant a aka. a * x.

Saves the resulting vector back into x. This is a Level 1 BLAS operation.

fn swap(&self, x: &mut SharedMemory<F>, y: &mut SharedMemory<F>) -> Result<(), Error>

Swapes the content of vector x and vector y.

Saves the resulting vector back into x. This is a Level 1 BLAS operation.