Module collenchyma::prelude

A module meant to be glob imported when using Collenchyma.

For instance:

use collenchyma::prelude::*;

This module contains several important traits that provide many of the convenience methods in Collenchyma, as well as most important types. Another type that is often needed but is likely to cause a name collision when imported is collenchyma::Error.

Reexports

pub use backend::*;

pub use device::{IDevice, DeviceType};

pub use hardware::{IHardware, HardwareType};

pub use framework::IFramework;

pub use memory::{IMemory, MemoryType};

pub use tensor::{SharedTensor, TensorDesc, ITensorDesc, IntoTensorDesc};

pub use frameworks::Native;