Module collenchyma::shared_memory

Provides the functionality for memory management accross devices.

A SharedMemory tracks the memory copies accross the devices of the Backend and manages

the location of these memory copies

the location of the latest memory copy and

the synchronisation of memory copies between devices

A memory copy represents one logical unit of data, which might me located at the host. The SharedMemory, tracks the location of the data blob accross the various devices that the backend might consist of. This allows us to run operations on various backends with the same data blob.

Examples

Create SharedMemory and fill it with some numbers:

#![feature(clone_from_slice)]
use collenchyma::framework::IFramework;
use collenchyma::frameworks::Native;
use collenchyma::shared_memory::SharedMemory;
// allocate memory
let native = Native::new();
let device = native.new_device(native.hardwares()).unwrap();
let shared_data = &mut SharedMemory::<i32>::new(&device, 5);
// fill memory with some numbers
let local_data = [0, 1, 2, 3, 4];
let data = shared_data.get_mut(&device).unwrap().as_mut_native().unwrap();
data.as_mut_slice().clone_from_slice(&local_data);

Structs

SharedMemory
Container that handles synchronization of Memory of type T.

Enums

Error
Errors than can occur when synchronizing memory.